Adaptando entidades

+++ src/main/java/upeu/edu/pe/lp/infrastructure/entity/ProductEntity.java
@@ -1,7 +1,6 @@
 package upeu.edu.pe.lp.infrastructure.entity;
 
 import jakarta.persistence.*;
-import lombok.Getter;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
@@ -17,7 +16,7 @@ public class ProductEntity {
     private String name;
     private String description;
     private String image;
-    private double price;
+    private BigDecimal price;
     
     private LocalDateTime dataCreated;
     private LocalDateTime dataUpdated;
@@ -69,11 +68,11 @@ public class ProductEntity {
         this.image = image;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
